backend/course_generator/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
aiohttp==3.9.1
tiktoken==0.5.2
pydantic==2.5.0
python-multipart==0.0.6
